Update gop.cpp
my understanding is that the output files should be able to  be binarily concatenated to become one bitstream, but no-annexb and the "lenx", pts and dts extra data breaks that.
maybe there're some reason to do so (apart from the requirement for the specifically designed tool and the process of mp4 muxing), and there's a specifically designed tool to mux it into mp4, but I prefer this way.
just concat them, and output is exactly the same as "raw" bitstream.

as of why not to force no-repeat-headers, well, it's not necessary, I presume.

and of course the compatibility to the tool is broken.