aq-mode 2523 and aq-bias-strength
AQ1: *joins the orgy*
Since I don't know if people will just keep assigning AQ2 variants to next and next numbers so I just use something big.
Unlike in amod's AQ5 (I don't know if it was in fact originated from amod or not), where they add aq-bias-strength as a important feature because dark bias in default situation was not punchy enough for AQ4, for AQ1 it's normally enough, so it's more like a testing/playing feature.
I was planning to use this param to set the bias_strength directly but gave up after I found that amod already has this param and it's more reasonable to use it as factor. (perhaps some scattered memories of it inspired me to reinvent this wheel)
2523=26*97+1
Bucause 2522 is a abandoned draft which only adds bias_strength * (1.f - modeTwoConst / (qp_adj * qp_adj)) to AQ1 code and the bitrate is skyrocketed.