remove b-adapt check for field option and move SEI param info
this check is incorrect (should be != 0 not == 0), since "field" feature is semifinished (assumed from // to do) and the interlace feature is experimental to begin with, and nobody cares about interlaced video encoding these days, there's no point correcting it.

also no point printing it in SEI param info if not used with interlace. because all it does it separate the fields according to interlace mode for you (and it's buggy), you can do it better yourself. (interlacing in HEVC is just separate fields encoded like a normal sequence with a SEI for each field indicating the order.)