decouple qp=4 from lossless mode
This can be considered as experimental, but it's also not at the same time.
The result will still be lossless.

QP will still affect compression ratio a bit, and QP=4 is not always the best.

Because the practical use of lossless mode is limited, and x264 lossless often have better compression ratio and faster decoding speed, I'm willing to consider this patch as experimental.