Errors: the HARD MODE
Inside me something's not right.
previously on numErrorsDuringEncoding:
move it (decl. and def.) to common.{h,cpp}, will cause multiple definition with multilib build, or if being inside X265_NS, the wrong one will be used in the final "return" (from the default bitdepth library, not the chosen bitdepth, I assume).
but with it in x265.cpp, the shared library cannot be built, now you can have a shared library that kills the whole process when error happens. (and of course static library will do the same)