if it ain't break, don't fix it
Don't change the default value of bScenecut, instead detect if we are using bframes at all.
Changing the bScenecut breaks hist-scenecut, although I'll probably never use it, better not let it happen.
Also the scenecut column in csv log is reading the bScenecut, but it was unchanged if there's not bframes (thus no flash detection).